+++ lib/src/clixon_datastore_write.c
@@ -483,7 +483,7 @@ text_modify(clicon_handle       h,
 		/* RFC 8040 4.6 PATCH:
 		 * If the target resource instance does not exist, the server MUST NOT create it. 
 		 */
-		if (netconf_data_missing(cbret, NULL,
+		if (netconf_data_missing(cbret, 
 	  "RFC 8040 4.6. PATCH: If the target resource instance does not exist, the server MUST NOT create it") < 0)
 		    goto done;
 		goto fail;
@@ -655,7 +655,7 @@ text_modify(clicon_handle       h,
 	    break;
 	case OP_DELETE:
 	    if (x0==NULL){
-		if (netconf_data_missing(cbret, NULL, "Data does not exist; cannot delete resource") < 0)
+		if (netconf_data_missing(cbret, "Data does not exist; cannot delete resource") < 0)
 		    goto done;
 		goto fail;
 	    }
@@ -676,7 +676,7 @@ text_modify(clicon_handle       h,
 		}
 		else {
 		    if (op == OP_DELETE){
-			if (netconf_data_missing(cbret, NULL, "Data does not exist; cannot delete resource") < 0)
+			if (netconf_data_missing(cbret, "Data does not exist; cannot delete resource") < 0)
 			    goto done;
 			goto fail;
 		    }
@@ -882,7 +882,7 @@ text_modify(clicon_handle       h,
 	    break;
 	case OP_DELETE:
 	    if (x0==NULL){
-		if (netconf_data_missing(cbret, NULL, "Data does not exist; cannot delete resource") < 0)
+		if (netconf_data_missing(cbret, "Data does not exist; cannot delete resource") < 0)
 		    goto done;
 		goto fail;
 	    }
@@ -1001,7 +1001,7 @@ text_modify_top(clicon_handle       h,
 	 I.e., curl -u andy:bar -sS -X DELETE http://localhost/restconf/data
       */
 	    case OP_DELETE:
-		if (netconf_data_missing(cbret, NULL, "Data does not exist; cannot delete resource") < 0)
+		if (netconf_data_missing(cbret, "Data does not exist; cannot delete resource") < 0)
 		    goto done;
 		goto fail;
 		break;